refactor: Move `Playroom` to sub-component folder
