+++ src/Playroom/Playroom/Playroom.css.ts
@@ -9,11 +9,11 @@ import {
   ANIMATION_TIMEOUT,
   toolbarItemCount,
   toolbarOpenSize,
-} from './constants';
+} from '../constants';
 
-import { toolbarItemSize } from './ToolbarItem/ToolbarItem.css';
-import { sprinkles, colorPaletteVars } from '../css/sprinkles.css';
-import { vars } from '../css/vars.css';
+import { sprinkles, colorPaletteVars } from '../../css/sprinkles.css';
+import { vars } from '../../css/vars.css';
+import { toolbarItemSize } from '../ToolbarItem/ToolbarItem.css';
 
 const MIN_HEIGHT = `${toolbarItemSize * toolbarItemCount}px`;
 const MIN_WIDTH = `${toolbarOpenSize + toolbarItemSize + 80}px`;
